skillDepartmentPlatformsCRUD
- Code Optimization

+++ src/test/java/com/testcases/CustomizationsTC.java
@@ -21,31 +21,82 @@ public class CustomizationsTC extends BaseTest {
 
 	@Test(priority = 1)
 	public void addSkill() {
-		customizationsPage = new CustomizationsPage();
-		SideMenuPage.sideMenuClick("Customizations");
-		SideMenuPage.subMenu("Skills");
-		customizationsPage.addSkill();
-		assertEquals(SideMenuPage.getToastMessage(), AppConstant.SKILL_ADD);
-		SideMenuPage.hiddenToastMessage();
+		skillDepartmentPlatformsCRUD("Customizations", "Skills", "add");
 	}
 
-	@Test(priority = 2, dependsOnMethods = "addSkill")
+	@Test(dependsOnMethods = "addSkill")
 	public void editSkill() {
-		customizationsPage = new CustomizationsPage();
-		SideMenuPage.sideMenuClick("Customizations");
-		SideMenuPage.subMenu("Skills");
-		customizationsPage.editSkill();
-		assertEquals(SideMenuPage.getToastMessage().trim(), AppConstant.SKILL_UPDATE);
-		SideMenuPage.hiddenToastMessage();
+		skillDepartmentPlatformsCRUD("Customizations", "Skills", "edit");
 	}
 
-	@Test(priority = 3, dependsOnMethods = "editSkill")
+	@Test(dependsOnMethods = "editSkill")
 	public void deleteSkill() {
+		skillDepartmentPlatformsCRUD("Customizations", "Skills", "delete");
+	}
+
+	@Test(priority = 2)
+	public void addDepartment() {
+		skillDepartmentPlatformsCRUD("Customizations", "Departments", "add");
+	}
+
+	@Test(dependsOnMethods = "addDepartment")
+	public void editDepartment() {
+		skillDepartmentPlatformsCRUD("Customizations", "Departments", "edit");
+	}
+
+	@Test(dependsOnMethods = "editDepartment")
+	public void deleteDepartment() {
+		skillDepartmentPlatformsCRUD("Customizations", "Departments", "delete");
+	}
+
+	@Test(priority = 3)
+	public void addPlatforms() {
+		skillDepartmentPlatformsCRUD("Customizations", "Platforms", "add");
+	}
+
+	@Test(dependsOnMethods = "addPlatforms")
+	public void editPlatforms() {
+		skillDepartmentPlatformsCRUD("Customizations", "Platforms", "edit");
+	}
+
+	@Test(dependsOnMethods = "editPlatforms")
+	public void deletePlatforms() {
+		skillDepartmentPlatformsCRUD("Customizations", "Platforms", "delete");
+	}
+
+	public void skillDepartmentPlatformsCRUD(String menu, String subMenu, String operation) {
 		customizationsPage = new CustomizationsPage();
-		SideMenuPage.sideMenuClick("Customizations");
-		SideMenuPage.subMenu("Skills");
-		customizationsPage.deleteSkill();
-		assertEquals(SideMenuPage.getToastMessage().trim(), AppConstant.SKILL_DELETE);
+		SideMenuPage.sideMenuClick(menu);
+		SideMenuPage.subMenu(subMenu);
+
+		if (operation.equals("add")) {
+			customizationsPage.addSkill();
+			if (subMenu.equals("Skills")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.SKILL_ADD);
+			} else if (subMenu.equals("Departments")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.DEPARTMENT_ADD);
+			} else if (subMenu.equals("Platforms")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.PLATFORM_ADD);
+			}
+		} else if (operation.equals("edit")) {
+			customizationsPage.editSkill();
+			if (subMenu.equals("Skills")) {
+				assertEquals(SideMenuPage.getToastMessage().trim(), AppConstant.SKILL_UPDATE);
+			} else if (subMenu.equals("Departments")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.DEPARTMENT_UPDATE);
+			} else if (subMenu.equals("Platforms")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.PLATFORM_UPDATE);
+			}
+		} else if (operation.equals("delete")) {
+			customizationsPage.deleteSkill();
+			if (subMenu.equals("Skills")) {
+				assertEquals(SideMenuPage.getToastMessage().trim(), AppConstant.SKILL_DELETE);
+			} else if (subMenu.equals("Departments")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.DEPARTMENT_DELETE);
+			} else if (subMenu.equals("Platforms")) {
+				assertEquals(SideMenuPage.getToastMessage(), AppConstant.PLATFORM_DELETE);
+			}
+		}
 		SideMenuPage.hiddenToastMessage();
 	}
 }
